Defer namespace modifications
We have a few core keywords that are also Redis commands, the
convenience of `$client->keys()` etc. is enough to justify
poking around with a UNITCHECK block to defer importing of
Net::Async::Redis::Commands entries to our namespace.

Local overrides are likewise runtime - `around` etc. happen
safely after all the code has been compiled - so this should
be enough to contain the complexity in a single

```
UNITCHECK { ->import }
```

and allow everything else to Just Work without any `CORE::`
prefixes.